+++ src/sdk_include/dolphin/ar.h
@@ -10,11 +10,11 @@ typedef void (*ARQCallback)(struct ARQRequest *);
 struct ARQRequest
 {
     /* 0x00 */ struct ARQRequest *next;
-    /* 0x04 */ u32 owner;
+    /* 0x04 */ uintptr_t owner;
     /* 0x08 */ u32 type;
     /* 0x0C */ u32 priority;
-    /* 0x10 */ u32 source;
-    /* 0x14 */ u32 dest;
+    /* 0x10 */ uintptr_t source;
+    /* 0x14 */ uintptr_t dest;
     /* 0x18 */ u32 length;
     /* 0x1C */ ARQCallback callback;
 };
@@ -53,9 +53,9 @@ u32 ARGetSize(void);
 // arq.c
 void ARQInit(void);
 void ARQReset(void);
-void ARQPostRequest(struct ARQRequest * request, u32 owner, u32 type, u32 priority, u32 source, u32 dest, u32 length, ARQCallback callback);
+void ARQPostRequest(struct ARQRequest * request, uintptr_t owner, u32 type, u32 priority, uintptr_t source, uintptr_t dest, u32 length, ARQCallback callback);
 void ARQRemoveRequest(struct ARQRequest * request);
-void ARQRemoveOwnerRequest(u32 owner);
+void ARQRemoveOwnerRequest(uintptr_t owner);
 void ARQFlushQueue(void);
 void ARQSetChunkSize(u32 size);
 u32 ARQGetChunkSize(void);
