+++ src/sysdolphin/baselib/devcom.c
@@ -98,7 +98,7 @@ static void HSD_DevComARAMCallback(ARQRequest* request)
     }
 
     if (aramDC->callback != NULL) {
-        aramDC->callback(aramDC->dcReq, (int) aramDC->args, buf,
+        aramDC->callback(aramDC->dcReq, (uintptr_t) aramDC->args, buf,
                          aramDC->cancelflag);
     }
 
@@ -137,7 +137,7 @@ void HSD_DevComARAMWakeUp(void)
     if (devComStatus[3] != NULL) {
         if (aramDC->cancelflag) {
             if (aramDC->callback != NULL) {
-                aramDC->callback(aramDC->dcReq, (s32) aramDC->args, NULL,
+                aramDC->callback(aramDC->dcReq, (uintptr_t) aramDC->args, NULL,
                                  true);
             }
             HSD_DevComUnlink(aramDC);
@@ -248,7 +248,7 @@ static void HSD_DevComDVDARAMEndCallback(ARQRequest* request)
 
     if (HSD_DevCom_804D77FC[i]->callback != NULL && HSD_DevCom_804D7804 == 0) {
         HSD_DevCom_804D77FC[i]->callback(
-            HSD_DevCom_804D77FC[i]->dcReq, (int) HSD_DevCom_804D77FC[i]->args,
+            HSD_DevCom_804D77FC[i]->dcReq, (uintptr_t) HSD_DevCom_804D77FC[i]->args,
             NULL, HSD_DevCom_804D77FC[i]->cancelflag);
     }
     HSD_DevComARAMCallback_inline(HSD_DevCom_804D77FC[i]);
@@ -275,7 +275,7 @@ static void HSD_DevComDVDMemCallback(s32 result, DVDFileInfo* unused)
         return;
     }
     if (dvdDC->callback != NULL && HSD_DevCom_804D7804 == 0) {
-        dvdDC->callback(dvdDC->dcReq, (int) dvdDC->args, NULL,
+        dvdDC->callback(dvdDC->dcReq, (uintptr_t) dvdDC->args, NULL,
                         dvdDC->cancelflag);
     }
     HSD_DevComUnlink(dvdDC);
@@ -306,7 +306,7 @@ static void HSD_DevComDVDCallback(s32 result, DVDFileInfo* unused)
         HSD_ASSERT(0x18C, dvdDC->size <= DEVCOM_BUF_SIZE);
         HSD_ASSERT(0x18D, dvdDC->callback);
         if (HSD_DevCom_804D7804 == 0) {
-            dvdDC->callback(dvdDC->dcReq, (s32) dvdDC->args,
+            dvdDC->callback(dvdDC->dcReq, (uintptr_t) dvdDC->args,
                             HSD_DevCom_804C6330_bufs[HSD_DevCom_804D77F6],
                             dvdDC->cancelflag);
         }
@@ -360,7 +360,7 @@ void HSD_DevComDVDWakeUp(void)
         if ((dvdDC = devComStatus[i])) {
             if (dvdDC->cancelflag) {
                 if (dvdDC->callback != NULL) {
-                    dvdDC->callback(dvdDC->dcReq, (s32) dvdDC->args, NULL,
+                    dvdDC->callback(dvdDC->dcReq, (uintptr_t) dvdDC->args, NULL,
                                     true);
                 }
                 HSD_DevComUnlink(dvdDC);
@@ -497,7 +497,7 @@ int HSD_DevComCancelEx(int dcReq, u32 flags, HSD_DevComCallback cb, void* args)
             dc->cancelflag = true;
         } else {
             if (dc->callback != NULL) {
-                dc->callback(dc->dcReq, (int) dc->args, NULL, true);
+                dc->callback(dc->dcReq, (uintptr_t) dc->args, NULL, true);
             }
             HSD_DevComUnlink(dc);
         }
